Completed Reflections
Created the infrastructure for Reflections and Refractions. Utilized the
reflect() function in the Vec3 class to reflect rays and added in
reflection and refraction depth counters in RayTracer.h. Simply just
reflected the ray across an intersected objects normal at the point of
intersection and raytraced the reflected ray. Changed the color at the
intersection point before applying lighting.